requirements.txt
# FastAPI and dependencies
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validation and serialization
pydantic[email]==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP client for external APIs
httpx==0.26.0
tenacity==8.2.3

# Caching
redis==5.0.1
hiredis==2.3.2

# Background tasks
celery==5.3.4
kombu==5.3.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx-mock==0.28.0
faker==22.1.0

# Code quality
ruff==0.1.11
black==23.12.1
mypy==1.8.0
types-redis==4.6.0.20240106
types-passlib==1.7.7.20240106

# External API integrations
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3.post1

# WebSocket support
websockets==12.0

# Monitoring and observability
prometheus-client==0.19.0
structlog==24.1.0

# PDF generation
reportlab==4.0.8
pypdf==3.17.4

# Excel export
openpyxl==3.1.2